Add files via upload
Multer is npm package which is use to store images and other files to server's disk. It doesn't store images on the database... wait then how we store our images on database ANSWER: using  "destination" and "filename" are the properties of multer.diskstorage in which we tell where to store the the image and what will be the name of the image file. Also we can control what type of file extension is allow to store like pdf, jpeg etc